Lógica de permissões 80% OK.

+++ src/main/java/com/betrybe/agrix/service/CropService.java
@@ -1,12 +1,11 @@
 package com.betrybe.agrix.service;
 
 import com.betrybe.agrix.entity.*;
-import com.betrybe.agrix.exception.CropNotFoundException;
-import com.betrybe.agrix.exception.FarmNotFoundException;
+import com.betrybe.agrix.exception.*;
 
 import com.betrybe.agrix.repository.*;
-import com.betrybe.agrix.utils.*;
 import org.springframework.beans.factory.annotation.*;
+import org.springframework.security.core.context.*;
 import org.springframework.stereotype.*;
 
 import java.time.*;
@@ -21,7 +20,6 @@ public class CropService {
 
   private final CropRepository cropRepository;
   private final FarmService farmService;
-
 
   /**
    * Instantiates a new Crop service.
@@ -43,20 +41,31 @@ public class CropService {
    * @return the crop
    * @throws FarmNotFoundException the farm not found exception
    */
-  public Crop createCropForFarm(Long farmId, Crop crop) throws FarmNotFoundException {
+  public Crop createCropForFarm(Long farmId, Crop crop) throws FarmNotFoundException, AccessDeniedException {
     Farm farm = farmService.findByFarmId(farmId);
-    crop.setFarmId(farm);
 
+    Person currentUser = (Person) SecurityContextHolder.getContext().getAuthentication().getPrincipal();
+    if (!farm.getPerson().equals(currentUser) && !hasRole("ADMIN") && !hasRole("MANAGER")) {
+      throw new AccessDeniedException("Você não tem permissão para criar plantações,fale com um responsável!");
+    }
+
+    crop.setFarmId(farm);
     return cropRepository.save(crop);
   }
 
   /**
-   * Find all crops list.
+   * Find all crops for the current user.
    *
-   * @return the list
+   * @return the list of crops
    */
-  public List<Crop> findAllCrops() {
-    return cropRepository.findAll();
+  public List<Crop> findAllCropsForCurrentUser() {
+    Person currentUser = (Person) SecurityContextHolder.getContext().getAuthentication().getPrincipal();
+
+    if (hasRole("ADMIN") || hasRole("MANAGER")) {
+      return cropRepository.findAll();
+    } else {
+      return cropRepository.findByFarmIdPerson(currentUser);
+    }
   }
 
   /**
@@ -66,35 +75,52 @@ public class CropService {
    * @return the crop
    * @throws CropNotFoundException the crop not found exception
    */
-  public Crop findByCropId(Long cropId) throws CropNotFoundException {
-    return cropRepository.findById(cropId)
+  public Crop findByCropId(Long cropId) throws CropNotFoundException, AccessDeniedException {
+    Crop crop = cropRepository.findById(cropId)
             .orElseThrow(CropNotFoundException::new);
+
+    Person currentUser = (Person) SecurityContextHolder.getContext().getAuthentication().getPrincipal();
+    if (!crop.getFarmId().getPerson().equals(currentUser) && !hasRole("ADMIN") && !hasRole("MANAGER")) {
+      throw new AccessDeniedException();
+    }
+
+    return crop;
   }
+
 
   /**
    * Find crop by farm id list.
    *
    * @param farmId the farm id
-   * @return the list
+   * @return the list of crops
    * @throws FarmNotFoundException the farm not found exception
    */
-  public List<Crop> findCropByFarmId(Long farmId) throws FarmNotFoundException {
+  public List<Crop> findCropByFarmId(Long farmId) throws FarmNotFoundException, AccessDeniedException {
     Farm farm = farmService.findByFarmId(farmId);
 
-    return findAllCrops().stream()
+
+    Person currentUser = (Person) SecurityContextHolder.getContext().getAuthentication().getPrincipal();
+    if (!farm.getPerson().equals(currentUser) && !hasRole("ADMIN") && !hasRole("MANAGER")) {
+      throw new AccessDeniedException("Você não tem permissão para acessar essa plantação !");
+    }
+
+    return cropRepository.findAll().stream()
             .filter(crop -> crop.getFarmId().equals(farm))
             .collect(Collectors.toList());
   }
 
   /**
-   * Find crops by search date list.
+   * Find crops by search date for current user.
    *
-   * @param start the start
-   * @param end   the end
-   * @return the list
+   * @param start the start date
+   * @param end   the end date
+   * @return the list of crops
    */
   public List<Crop> findCropsBySearchDate(LocalDate start, LocalDate end) {
-    return cropRepository.findAllByHarvestDateBetween(start, end);
+    Person currentUser = (Person) SecurityContextHolder.getContext().getAuthentication().getPrincipal();
+    return cropRepository.findAllByHarvestDateBetween(start, end).stream()
+            .filter(crop -> crop.getFarmId().getPerson().equals(currentUser) || hasRole("ADMIN") || hasRole("MANAGER"))
+            .collect(Collectors.toList());
   }
 
   /**
@@ -104,9 +130,15 @@ public class CropService {
    * @return the crop
    * @throws CropNotFoundException the crop not found exception
    */
-  public Crop updatedCrop(Crop crop) throws CropNotFoundException {
+  public Crop updatedCrop(Crop crop) throws CropNotFoundException, AccessDeniedException {
     Crop existingCrop = cropRepository.findById(crop.getId())
             .orElseThrow(CropNotFoundException::new);
+
+    // Verifica se a crop pertence ao usuário autenticado ou se o usuário tem a role de ADMIN ou MANAGER
+    Person currentUser = (Person) SecurityContextHolder.getContext().getAuthentication().getPrincipal();
+    if (!existingCrop.getFarmId().getPerson().equals(currentUser) && !hasRole("ADMIN")) {
+      throw new AccessDeniedException("Você não tem acesso para editar essa plantação, peça ao ADMIN!");
+    }
 
     existingCrop.setName(crop.getName());
     existingCrop.setPlantedArea(crop.getPlantedArea());
@@ -119,13 +151,24 @@ public class CropService {
    * Delete crop string.
    *
    * @param id the id
-   * @return the string
    * @throws CropNotFoundException the crop not found exception
    */
-  public String deleteCrop(Long id) throws CropNotFoundException {
+  public void deleteCrop(Long id) throws CropNotFoundException, AccessDeniedException {
     Crop existingCrop = cropRepository.findById(id)
             .orElseThrow(CropNotFoundException::new);
+
+    // Verifica se a crop pertence ao usuário autenticado
+    Person currentUser = (Person) SecurityContextHolder.getContext().getAuthentication().getPrincipal();
+    if (!existingCrop.getFarmId().getPerson().equals(currentUser) && !hasRole("ADMIN")) {
+      throw new AccessDeniedException("Você não tem acesso para excluir essa plantação, peça ao ADMIN!");
+    }
+
     cropRepository.delete(existingCrop);
-    return MessageUtil.CROP_DELETED;
+  }
+
+  private boolean hasRole(String role) {
+    return SecurityContextHolder.getContext().getAuthentication()
+            .getAuthorities().stream()
+            .anyMatch(grantedAuthority -> grantedAuthority.getAuthority().equals(role));
   }
 }
